low-risk-fix: 完全修复审计低风险项 L1-L8 (SMAP/LBA48/栈金丝雀/LFN/构建校验)

+++ include/kernel/ata.h
@@ -22,8 +22,9 @@ bool ata_read_sectors(uint32_t lba, uint8_t count, void *buf);
 /* PIO 写 count 个扇区（写后执行 FLUSH CACHE 确保落盘）；成功返回 true */
 bool ata_write_sectors(uint32_t lba, uint8_t count, const void *buf);
 
-/* 磁盘总扇区数（IDENTIFY 报告的 LBA28 容量） */
-uint32_t ata_total_sectors(void);
+/* 磁盘总扇区数（L2 修复：优先 LBA48 容量 word100-103，不支持时回退 LBA28
+ * word60-61；>128GB 盘不再截断为 LBA28 上限） */
+uint64_t ata_total_sectors(void);
 
 /* 启动内核磁盘服务任务（拥有 DISK_PORT，处理 IPC 读请求） */
 void disk_srv_start(void);
